+++ core/test/src/main/java/com/sidgowda/pawcalc/test/TestModule.kt
@@ -1,9 +1,15 @@
 package com.sidgowda.pawcalc.test
 
+import com.sidgowda.pawcalc.data.dogs.datasource.DogsDataSource
+import com.sidgowda.pawcalc.data.dogs.datasource.DogsMemoryDataSource
+import com.sidgowda.pawcalc.data.dogs.repo.DogsRepo
+import com.sidgowda.pawcalc.data.dogs.repo.DogsRepoImpl
 import com.sidgowda.pawcalc.data.modules.DispatchersModule
+import com.sidgowda.pawcalc.data.modules.DogsDataModule
 import com.sidgowda.pawcalc.data.modules.SettingsModule
 import com.sidgowda.pawcalc.data.settings.datasource.SettingsDataSource
 import com.sidgowda.pawcalc.data.settings.repo.SettingsRepo
+import com.sidgowda.pawcalc.test.fakes.FakeDogsDiskDataSource
 import com.sidgowda.pawcalc.test.fakes.FakeSettingsDataSource
 import com.sidgowda.pawcalc.test.fakes.FakeSettingsRepo
 import dagger.Module
@@ -16,6 +22,41 @@ import javax.inject.Named
 import javax.inject.Singleton
 
 object TestModule {
+
+    @Module
+    @TestInstallIn(
+        components = [SingletonComponent::class],
+        replaces = [DogsDataModule::class]
+    )
+    object TestDogsDataModule {
+
+        @Named("memory")
+        @Singleton
+        @Provides
+        fun providesMemoryDogsDataSource(
+            settingsDataSource: SettingsDataSource,
+            @Named("computation") computationDispatcher: CoroutineDispatcher
+        ): DogsDataSource {
+            return DogsMemoryDataSource(settingsDataSource, computationDispatcher)
+        }
+
+        @Named("disk")
+        @Singleton
+        @Provides
+        fun providesDiskDogsDataSource(): DogsDataSource {
+            return FakeDogsDiskDataSource()
+        }
+
+        @Singleton
+        @Provides
+        fun providesDogRepo(
+            @Named("memory") memoryDataSource: DogsDataSource,
+            @Named("disk") diskDataSource: DogsDataSource,
+            @Named("computation") computationDispatcher: CoroutineDispatcher
+        ): DogsRepo {
+            return DogsRepoImpl(memoryDataSource, diskDataSource, computationDispatcher)
+        }
+    }
 
     @Module
     @TestInstallIn(
